Changed function names from camelCase to underscores.
Reason: Consistency with the standard library, and it often looks better. There's an inconsistency already between OpenCV, Eigen and std::, so it's best if we just stick to the std:: convention.
The rest of the code will be changed step by step. Will change variable names to underscores as well.